+++ opwen_email_server/config.py
@@ -34,7 +34,7 @@ CONTAINER_AUTH = f'clientsauth{resource_suffix}'
 REGISTER_CLIENT_QUEUE = f'register{resource_suffix}'
 INBOUND_STORE_QUEUE = f'inbound{resource_suffix}'
 WRITTEN_STORE_QUEUE = f'written{resource_suffix}'
-PROCESS_SERVICE_QUEUE = f'service{resource_suffix}'
+PROCESS_ECHO_SERVICE_QUEUE = f'echo{resource_suffix}'
 SEND_QUEUE = f'send{resource_suffix}'
 MAILBOX_RECEIVED_QUEUE = f'mailboxreceived{resource_suffix}'
 MAILBOX_SENT_QUEUE = f'mailboxsent{resource_suffix}'
