+++ opwen_email_server/integration/connexion.py
@@ -18,7 +18,6 @@ from opwen_email_server.integration.azure import get_pending_storage
 from opwen_email_server.integration.azure import get_raw_email_storage
 from opwen_email_server.integration.azure import get_user_storage
 from opwen_email_server.integration.celery import inbound_store
-from opwen_email_server.integration.celery import process_service
 from opwen_email_server.integration.celery import register_client
 from opwen_email_server.integration.celery import written_store
 from opwen_email_server.services.auth import AnyOfBasicAuth
@@ -31,12 +30,6 @@ email_receive = ReceiveInboundEmail(
     auth=get_auth(),
     raw_email_storage=get_raw_email_storage(),
     next_task=inbound_store.delay,
-)
-
-service_request_email_receive = ReceiveInboundEmail(
-    auth=get_no_auth(),
-    raw_email_storage=get_raw_email_storage(),
-    next_task=process_service,
 )
 
 client_write = UploadClientEmails(
